turned value into float and added decimal place, now shows as currency

+++ db/penny_wise.sql
@@ -19,7 +19,7 @@ CREATE TABLE transactions
   id SERIAL8 primary key,
   merchant_id INT8 references merchants(id),
   category_id INT8 references categories(id),
-  value INT8,
+  value FLOAT,
   item VARCHAR(255) not null,
   transaction_date date
 );
